@@ -9,17 +9,18 @@
             "email": "bogdan.cismariu@gmail.com"
         }
     ],
-    "minimum-stability": "stable",
+    "minimum-stability": "dev",
+    "prefer-stable": true,
     "require": {
-        "php": "^7.2.5",
-        "illuminate/support": "^7.0",
-        "illuminate/database": "^7.0",
-        "illuminate/contracts": "^7.0"
+        "php": "^7.3",
+        "illuminate/support": "^8.0",
+        "illuminate/database": "^8.0",
+        "illuminate/contracts": "^8.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^8.0",
-        "orchestra/testbench": "^5.0",
-        "orchestra/database": "^5.0"
+        "phpunit/phpunit": "^9.0",
+        "orchestra/testbench": "^6.0",
+        "orchestra/database": "^6.0"
     },
     "autoload": {
         "psr-4": {
